added CDI producer for Context Util, moved it into Payara API
