+++ nucleus/common/internal-api/pom.xml
@@ -102,6 +102,11 @@
             <artifactId>class-model</artifactId>
         </dependency>
         <dependency>
+            <groupId>fish.payara.api</groupId>
+            <artifactId>payara-api</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
             <groupId>fish.payara.server.internal.common</groupId>
             <artifactId>glassfish-api</artifactId>
             <version>${project.version}</version>
